+++ react/vitest.config.ts
@@ -10,7 +10,7 @@ import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
     test: {
-        include: ['test/**/*.test.ts'],
+        include: ['test/**/*.test.{ts,tsx}'],
         clearMocks: true,
         restoreMocks: true,
         unstubGlobals: true